+++ components/retro-go/rg_sdcard.h
@@ -3,19 +3,19 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#define RG_BASE_PATH           "/sd"
-#define RG_BASE_PATH_CACHE     RG_BASE_PATH "/retro-go/cache"
-#define RG_BASE_PATH_ROMART    RG_BASE_PATH "/romart"
-#define RG_BASE_PATH_ROMS      RG_BASE_PATH "/roms"
-#define RG_BASE_PATH_SYSTEM    RG_BASE_PATH "/retro-go/system"
-#define RG_BASE_PATH_TEMP      RG_BASE_PATH "/retro-go/temp"
+#define RG_ROOT_PATH            "/sd"
+#define RG_BASE_PATH            RG_ROOT_PATH "/retro-go"
+#define RG_BASE_PATH_CACHE      RG_BASE_PATH "/cache"
+#define RG_BASE_PATH_CONFIG     RG_BASE_PATH "/config"
+#define RG_BASE_PATH_COVERS     RG_ROOT_PATH "/romart"
+#define RG_BASE_PATH_ROMS       RG_ROOT_PATH "/roms"
+#define RG_BASE_PATH_SAVES      RG_BASE_PATH "/saves"
+#define RG_BASE_PATH_SYSTEM     RG_BASE_PATH "/system"
+#define RG_BASE_PATH_THEMES     RG_BASE_PATH "/themes"
 
 #ifdef RG_TARGET_ODROID_GO
-#define RG_BASE_PATH_CONFIG    RG_BASE_PATH "/odroid"
-#define RG_BASE_PATH_SAVES     RG_BASE_PATH "/odroid/data"
-#else
-#define RG_BASE_PATH_CONFIG    RG_BASE_PATH "/retro-go/config"
-#define RG_BASE_PATH_SAVES     RG_BASE_PATH "/retro-go/saves"
+#undef RG_BASE_PATH_SAVES
+#define RG_BASE_PATH_SAVES     RG_ROOT_PATH "/odroid/data"
 #endif
 
 bool rg_sdcard_mount(void);
@@ -29,4 +29,4 @@ bool rg_mkdir(const char *dir);
 // These are similar to libgen's but they never modify the source
 const char *rg_dirname(const char *path);
 const char *rg_basename(const char *path);
-const char* rg_extension(const char *path);
+const char *rg_extension(const char *path);